[Feature:Forum] Add ability to clear live chat messages (#12008)
### Why is this Change Important & Necessary?
Closes #11998. If an instructor wants to clear a chat for any reason,
previously the only option was to delete the chat and create a new one.

### What is the New Behavior?
This feature adds a new button to the Chatroom Row component which
clears the associated live chat, with a dialog to confirm and a new
websocket action which deletes a message. The clear endpoint sends the
websocket request once for each message it is deleting. A flag is
created for each message called is_deleted which will allow us to query
for deleted messages if we want to see them.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Run `submitty_install` on this branch, and create a live chat. The clear
button should be visible, and having a student client open within the
live chat should allow you to see that when an instructor clicks clear,
all messages disappear dynamically.

<img width="3838" height="1915" alt="image"
src="https://github.com/user-attachments/assets/71b9f232-3dfc-43dd-a06c-60da24baded1"
/>

### Automated Testing & Documentation
New tests were written for this feature, including one for the websocket
feature.

### Other information
This feature has migrations. A future PR should be opened to be able to
delete individual messages and a panel should be added to view/restore
deleted messages.
Relies on PR #11863

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>