[Bugfix:Developer] Fix Sidebar Cypress Test (#12078)
### Why is this Change Important & Necessary?
Currently, the sidebar cypress test fails on main due to a cypress
version mismatch with an external plugin.
Before, in this situation we were using 'attachFile' which is from an
external plugin.

### What is the New Behavior?
Now, we are using the built-in function 'selectFile' which uses a
different location, but it's easily changed.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Mainly, just review the code, and check that the tests are now passing
in CI.

### Other information
This will fix one of the many failing tests in Main.