[Bugfix:TAGrading] Show autograding stats before TAGrading (#12085)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Currently if a gradeable has autograding but no tagrading or autograding
with unfinished tagrading, student that do not have tagrading data will
not be included in autograding statistics.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Any student with autograding outside of the null section will be
included in autograding statistics.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Submit to a gradeable with autograding on main (I used tutorial 01)
2. Notice that the autograding statistics do not change
3. View on branch

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->