[Feature:Notifications] Mark Seen for Home Page (#12007)
### Why is this Change Important & Necessary?
The home page does not have any way to mark notifications as seen,
besides clicking them and going to the link. This can be cumbersome,
especially with the new gradeable notifications whenever submissions
open.

### What is the New Behavior?
Added an open mail icon next to each individual notification that is
unseen. Clicking this icon marks that individual notification as seen,
and does not redirect the user to its source. This icon is not visible
on seen notifications.

Since I needed to add a new link inside what was previously an anchor
acting as a container, I refactored each notification into a div with
two separate anchor elements. That’s why the diff appears so messy.

**Mark all as seen:**
This PR also includes the `Mark all as seen` button and functionality,
which is all commented out for now. The button functions as follows:

> Added a `Mark all as seen` button in the top right of the
notifications panel. Clicking this button marks ALL of the user's
notifications as seen, not just the ones displayed on the page. The
request to the backend is not made when the user has no unread
notifications in any of their 10 most recent from each of their active
courses. I chose the blue color for this button while keeping `Show All`
white because users will be clicking it frequently.

<img width="1620" height="742" alt="image"
src="https://github.com/user-attachments/assets/1c44bced-0adf-4f38-a7e0-d8a61351e32a"
/>

### What steps should a reviewer take to reproduce or test the bug or
new feature?
[OPTIONAL] Run `recreate_sample_course` to have many notifications for
`Submissions Open`.
Verify that Individual and Mark All buttons function correctly.

### Automated Testing & Documentation
This feature is not yet sufficiently tested with E2E tests. I will make
a follow-up PR for this.

### Other information
This is not a breaking change.