[Bugfix:Autograding] system calls on pdf_notes_page gradeables (#12077)
### Why is this Change Important & Necessary?
The list of allowed systems calls for pdflatex on the notes page
gradeables needs to be updated.

### What is the New Behavior?
Before the change, the pdf in testcase #2 containing the instructor
'exam' with the student notes attached was not viewable.
The pdf was not produced/was corrupted because a system call was failing
during one of the pdf processing steps.
After the change, the pdf is viewable.

### What steps should a reviewer take to reproduce or test the bug or
new feature?

### Automated Testing & Documentation
We should continue to add more automated testing of these gradeables.  

### Other information

Co-authored-by: Barb Cutler <Barb Cutler>