[Bugfix:Forum] Fix Clear Messages Socket (#12072)
### Why is this Change Important & Necessary?
The live chat clearing test module of `live_chat.spec.js` is failing on
CI. This failure was introduced in commit b66259f of #12008 when I
merged main into it. There were other live lecture chat features (namely
#11863) introduced in this merge, however, I could not locate a failing
test for the clear socket on any of them. NOTE: The clear feature worked
as expected up until that commit.

### What is the New Behavior?
N / A. I assume the change has to come from `ChatroomController.php`. It
is not the Cypress test that is broken, but rather the websocket for
live chat clearing.

### How to Test
Delete all of the tests in `live_chat_spec.js` except for the bottom so
you don't have to wait 5 minutes for it to pass.

### Other information
This is not a breaking change. This PR is currently marked as a draft
since I have been unable to come up with a fix. The current diff was
from when I thought the issue was timing before discovering it was
unrelated to Cypress.

Error on 10/03/2025 11:30:00
<img width="1827" height="261" alt="image"
src="https://github.com/user-attachments/assets/0e80361e-4077-4862-a8eb-4b61a0896e01"
/>
Looks promising, will continue to investigate.

Found another PR that touches live chat sockets:
#11634 (Add Verification of WebSocket Pages) - a lot of types were
changed. Seems like a small bug was introduced that broke websockets for
in #12008 (Clear Button for Live Chat) - still investigating, but
narrowing it down.

---------

Co-authored-by: Alex Lavallee <73203142+lavalleeale@users.noreply.github.com>
Co-authored-by: Alex Lavallee <alex@lavallee.one>