[Refactor:Notifications] Individual Notification Component (#12137)
### Why is this Change Important & Necessary?
`HomePage.vue` has a component, `AllNotificationsDisplay.vue`, that
functions as the entire notification panel for that page.

As more things are added to this file (such as new buttons when we
implement it for the Course Notifications page), it is becoming large
(~400 lines), and we losing the benefits of the component-based approach
that Vue offers.

### What is the New Behavior?
Essentially, this PR splits up the `AllNotificationsDisplay.vue` file.
Instead of looping through and displaying divs defined in that file for
each individual notification, it now loops through and displays a
component, `Notification`, for each notification. All styling and
notification-specific logic has been moved to the new file,
`Notification.vue`, dividing the original file into two ~200 line files.

This change just abstracts logic to improve maintainability. This may
help visualize the it (think of these are Vue components):

Before:
```
HomePage/
├── AllNotificationsDisplay (~400 lines)
```

After:
```
HomePage/
├── AllNotificationsDisplay (142 lines)
       ├── Notification (196 lines)
```

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Ensure that the functionality of the notifications panel on the home
page is unchanged.

### Automated Testing & Documentation
There are no automated tests for the notifications system, see #11908.

### Other information
This is not a breaking change. I intend to create a follow-up PR to this
to convert the Course Notifications page. By breaking the existing code
into multiple components, it will be easier to migrate, and maintain.

There is potential to add even more components beyond this, such as one
for the header, but these are subject for discussion. I think the
individual notifications is the most important for now.