[Feature:TAGrading] Make active graders an outline (#12128)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Active grader indicators were full red circles, which made it impossible
to see the status of the component outside of whether or not it is being
graded.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
The new behavior is that whether or not a component is being graded is
represented by a red outline on the component circle.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Start actively grading a component
2. View the grading details view on both main and the PR

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->