[Bugfix:Developer] Fix Ansible install in CI (#12153)
### Why is this Change Important & Necessary?
Currently the Ansible install tests are failing because the autograder
services can't start because of too much disk space.

### What is the New Behavior?
The same folders that are removed in the normal CI procedures are now
removed for Ansible install as well.
The tests should now run and pass.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
See the tests now pass in CI.

### Other information
We should investigate more options to reduce space, because it looks
like it's ~40+ gb.