[Bugfix:InstructorUI] Fix Course Materials Prefix Collisions (#12208)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Fixes #12203.
As detailed by the issue itself, there is a significant flaw in the
implementation of course materials file deletion. In the current
production version, if there is a directory which shares the same prefix
as another directory, such as `tests/test0` and `test/test01`, the
deletion of the `test01` subdirectory would (assuming no other
directories in `test`) cascade to delete `tests` and all of its
subdirectories. This is due to an improper usage of `str_starts_with` in
the `CourseMaterialsController`.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
The error has been corrected by enforcing a strict directory boundary to
avoid prefix collisions.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Checkout the `main` branch. As an instructor, create directories in a
course's course materials such that the exploit is reproducible, as in
`tests/test0` and `test/test01`, and then try deleting (in separate
attempts) the top directory, immediate subdirectory, or file with
respect to the directory that has the shorter handle. Notice the error.
Repeat this process with this branch (`course-materials-edit-bugfix`)
and notice that the behavior is as intended.