[Bugfix:Developer] Fix Registration Test (#12215)
### Why is this Change Important & Necessary?
CI is failing due to `registration.spec.js`. This was caused in #12154
from nesting a `<button>` inside of an `<h1>` element, so when Cypress
went to assert that the header said "My Courses", it returned "My
Courses &nbsp;" instead, causing the test to fail.

### What is the New Behavior?
Take the `<button>` out of the `<h1>`, and adjust the CSS to remain the
same.

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>