Working Mark Seen button
The user can now save their changes, but the page does not yet dynamically update.

REMAINING TO-DO: Add an index to the notifications table for userid WHERE seen_at is null. This will speed up the counts (I think?)

Dynamically update the home page with the new list. We should just need to pass back the courses that the user selected and remove any that match from the dynamic list. This will be far better than a page reload, and shouldn't be difficult.