[Feature:Notifications] Home Page Additional Unseen Count (#12227)
### Why is this Change Important & Necessary?
The home page notifications panel only displays the 10 most recent
notifications from each of the user's courses. If the user has 11 unseen
notifications in a course, they will not be able to see the last one on
the home page. Additionally, the user could have more than 10 unseen
notifications between all of their courses, and they would not be able
to see all of them, and there is no indicator as to how many they have
total.

### What is the New Behavior?
Added a `<p>` beneath the displayed notifications in the `Unseen Only`
mode that appears when the user has more than 10 unseen notifications
and displays the total count of unseen notifications - 10 (the ones
already displayed).

#### Light Mode
<img width="717" height="94" alt="image"
src="https://github.com/user-attachments/assets/3961ad96-c12d-4ab6-89ec-51c4757094c8"
/>

#### Dark Mode
<img width="717" height="94" alt="image"
src="https://github.com/user-attachments/assets/75e362d0-99e3-4d4a-a970-1efea015a89e"
/>

#### Cases
1. When there are additional unseen notifications in the front-end, but
they are not currently displayed.
2. There are additional unseen notifications, but they are not able to
be brought to the front-end and displayed.
**When no unseen notifications are displayed, but there are older
ones:**
<img width="1538" height="250" alt="image"
src="https://github.com/user-attachments/assets/2b761f87-380f-4008-a258-29b0947d8542"
/>

#### Implementation
When we grab the notifications for the home page, we loop through and
query each of the course databases, appending the results to an array.
This array is then serialized (to add extra data) when it is sent to
Vue.

When getting the total unseen count, I wanted to avoid the following
complications:
1. Messing up the serialization of the Notifications array.
2. Looping through the course databases two different times.

My solution was to change the original function name,
`getAllRecentNotifications` to `getHomeNotificationData`, pass in
`unseen_count` as reference, and add each course's unseen count to it in
the loop. I am unsure if this was the best solution, but at least it
avoids another loop.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Play around with the notifications panel. Test bulk mark seen,
individual mark seen, when the user has less than 10 unseen
notifications, and any other possible edge cases. If you have any
questions about my implementation, don't hesitate to ping me.

Additionally, ensure the course notifications page has unchanged
functionality.

### Automated Testing & Documentation
The entire notification system does not yet have adequate e2e testing
(see https://github.com/Submitty/Submitty/issues/11908). The
notifications UI is nearing completion, so perhaps this issue can be
addressed soon.

### Other information
This is not a breaking change, but it is on the home page so if anything
were to go wrong, it users would not be able to access the site.