[Bugfix:TAGrading] Prevent Credit Stealing (#12228)
### Why is this Change Important & Necessary?
Fixes #12223
TAs who have not made any modifications to a grading component are
receiving credit for the question just for navigating past it. This is
skewing the TA grading statistics, making it difficult for instructors
to keep of track of the progress of individual graders.

### What is the New Behavior?
The issue was caused by the `score` and `comment` values in the
`getGradedComponentFromDOM()` function. Although the function
initialized them to default values (0.0 and ''), those defaults were
never reapplied when the DOM lookup failed. When the inputs were missing
or empty, `scoreInput.val()` returned undefined, which led to
`parseFloat(undefined)` producing NaN, and the textarea lookup returning
null for comment.

The fix was to explicitly reset these values to their defaults whenever
the DOM returned NaN, undefined, or null, ensuring `score` always
resolves to 0.0 and `comment` always resolves to an empty string.

The above issue was causing `gradedComponentsEqual()` to return `false`
since it was checking the expected `0.0` and `''` against NAN and NULL,
in turn allowing graders who were on the page component but not making
any changes to receive/"steal" credit.

Below are the results of `gradedComponentsEqual()` from before and after
the fix when the user opens and closes a rubric component 2x.
#### Before
<img width="1229" height="275" alt="image"
src="https://github.com/user-attachments/assets/b981cdd4-a4fd-495d-a600-eaae3305e8cb"
/>

#### After
<img width="1229" height="275" alt="image"
src="https://github.com/user-attachments/assets/0ef09ffb-c48a-41d7-859e-f64e8d9c7323"
/>

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Replicate #12223 on main. You should be able to steal credit from
opening and closing the component 2x, as well as by leaving the
component open, and navigating to the next student.

### Automated Testing & Documentation
I don't believe we have tests for crediting graders, perhaps this should
be added to `rubric_grading.spec.js` in a future PR.

### Other information
This is not a breaking change