[Bugfix:Notifications] Misc Notification UI Bugfixes (#12233)
### Why is this Change Important & Necessary?
There are several small bugs and code convention errors in the
notification UI, mostly introduced by #12227. If necessary, this PR can
be split into multiple smaller ones, but since all changes are minor and
primarily contained in `NotificationsDisplay.vue`, I figured I would
just address all of them in a single PR.

#### UI  Issues
1. **Unseen notification count displayed when 0**
In the original PR, #12227, I did not have a condition for when older
notifications is equal to 0. This meant if the user had an unseen
notification displayed, the older notifications count would always be
visible, even when it was 0.

2. **"No unseen notifications" not appearing dynamically when course
notifications are cleared**
If the user marks an individual course notification as seen and then
clears all of them, the "No unseen notifications" text would not appear
until after a reload. This is because the course notification's prop for
unseenCount is set to 0, meaning it becomes negative if notifications
are marked seen, and thus preventing the text from displaying.

#### Developer Convention Issues
3. **Using Incorrect Notification Arrays**
`NotificationsDisplay.vue` contains 3 notification lists:
- `localNotifications` - all notifications sent from the backend
- `filteredNotifications` - filtered version of `localNotifications`
that is either unseen only, or all based on local storage variable
- `visibleNotifications` - filtered version of `filteredNotifications` -
all for courses, 10 most recent for home page.
When referring to displayed notifications, you should use
`visibleNotifications`. When manipulating the data in notifications
(i.e., mark seen), you should use `localNotifications`, since the
updates will propogate to the other arrays from there.

The logic introduced in #12227 uses `filteredNotifications`. I don't
think this affects the functionality, but it is still not the correct
convention.

4. **Inconsistent Logic**
The individual mark seen logic does not match the mark all as seen
logic.

5. **Using `semester` instead of `term`**
Submitty uses term, not semester. In #11914, the first PR for the new
notifications UI, I used semester. In more recent PRs such as #12219, I
used the correct Submitty convention, term, leading to code
inconsistencies.

### What is the New Behavior?
1. **Unseen notification count displayed when 0**
Added a condition to ensure the count is not displayed when it is 0.
#### Before
<img width="1541" height="362" alt="Screenshot from 2025-12-01 15-21-14"
src="https://github.com/user-attachments/assets/3e15c3b5-aaa9-49b5-9a41-1eecc3e75eaa"
/>

#### After
<img width="1541" height="362" alt="Screenshot from 2025-12-01 15-24-13"
src="https://github.com/user-attachments/assets/fb7494b3-e422-49b8-b303-e88c1a06274b"
/>

2. **"No unseen notifications" not appearing dynamically when course
notifications are cleared**
Rather than checking if `unseenCount` is equal to 0, check if
`unseenCount` is <= 0.

3. **Using Incorrect Notification Arrays**
Use `visibleNotifications` for the unseen count logic instead of
`filteredNotifications`. Move arrays to be next to each other and adjust
comments to make it more apparent which one should be used.

4. **Inconsistent Logic**
Use a for loop for both individual mark seen and mark all as seen.

5. **Using `semester` instead of `term`**
Changed all instances of `semester` in the notification UI to `term`.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
Play around with the notifications UI and ensure there are no other odd
behaviors.

### Other information
This is not a breaking change.

#### Minor Bug Discovery
In my testing, I discovered a possible race condition with clearing
individual notifications. When the user has many notifications and
attempts to rapidly clear them one at a time, it is possible for a
notification to appear as seen, but persist as the 11th notification in
the unseen only panel.
<img width="1541" height="362" alt="Screenshot from 2025-12-01 17-16-42"
src="https://github.com/user-attachments/assets/78b81db5-724b-47fc-9e31-a7f801fd1fbb"
/>
This seems to occur when there is a high volume of mark_seen network
requests, causing them to return out of order. The solution would be to
block each individual mark seen button until each request is completed,
but I believe that would be too much overhead for such a minor issue.

This should not occur often, as users with many notifications should
clear them using `Mark as seen`. Additionally, the issue can be solved
with a simple reload, and does not occur if the user is using the
notification UI as intended / not individually clearing 10+
notifications at a time.