[Refactor:CourseMaterials] Phpstan 2.1.32 compatibility (#12150)
### Why is this Change Important & Necessary?
<!-- Include any GitHub issue that is fixed/closed using "Fixes
#<number>" or "Closes #<number>" syntax.
Alternately write "Partially addresses #<number>" or "Related to
#<number>" as appropriate. -->
Before we can merge #12093 (superseded by #12260), we need to make sure
that all of our code is compatible with the new version of phpstan, and
this was the one place that our code has started failing.

### What is the New Behavior?
<!-- Include before & after screenshots/videos if the user interface has
changed. -->
Phpstan 2.1.32 now passes after this change.

### What steps should a reviewer take to reproduce or test the bug or
new feature?
1. Checkout #12260 and run `composer install` and then phpstan and view
that 1 error is detected
2. Run `git diff main phpstan-prep | git apply` to get the changes from
this branch
3. Run phpstan again and it shall pass

### Automated Testing & Documentation
<!-- Is this feature sufficiently tested by unit tests and end-to-end
tests?
If this PR does not add/update the necessary automated tests, write a
new GitHub issue and link it below.
Is this feature sufficiently documented on submitty.org?
Link related PRs or new GitHub issue to update documentation. -->

### Other information
<!-- Is this a breaking change?  
Does this PR include migrations to update existing installations?  
Are there security concerns with this PR? -->

---------

Co-authored-by: Ryan Styron <ryanstyron1220@gmail.com>